requirements.txt
# Plant Village Disease Classification - Dependencies
# Compatible with Python 3.10

# Core ML libraries
tensorflow==2.13.0
numpy>=1.21.0,<2.0.0
pandas>=1.3.0

# Image processing
opencv-python-headless>=4.5.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0

# Streamlit web framework
streamlit>=1.28.0

# FastAPI for REST API
fastapi>=0.100.0,<0.110.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Fix protobuf compatibility
protobuf>=3.19.0,<4.0.0
